[WIP] Trying to extract contours from 2D images
Just trying the active contour feature of scikit-image to
extract the border of the limbs data of the TIF file.

I've been tweaking things but it is not really working
as of now. I need to have the intuition and to understand
the properties of snakes. 🐍

I'll try to test canny edge detector which are also
available in scikit-image.