[WIP] Trying sobel filter + flooding
This adds some fixtures to try to extract contours using
sobol filter and some flooding.

In this context, the main parameter one can tune is the
tolerance.

Results are quite noisy, we might want to use something then
to extract the contour. Maybe the active contour technique
might work nicely in this context.

Next:
 - try to use some active contour techniques to extract
 the shape of the embryo
 - directly try marching cubes techniques