[WIP] Marching Cubes + Morphological AC
Here the implementation of Marching Cubes from
sklearn is used.

The two methods (lorenser, leviner) are tested.
They give quasi-identical results on the meshes.

The hardest is to tune the threshold, this has
to be done manually, unfortunately.

Morphological active contour techniques from [1]
are tested and they give _pretty_ good results.

Though I haven't enough memory on my machine to
run it on the completed resolution (which is 512³),
it gives quite a reasonably good segmentation on a
lower resolution (256³).

The implementation really can't be improved.
The intensive memory usage come from having two 512³
arrays of uint8 (approx 8GioB).

I ran the algo on several halves, I'll try to see
if it is possible to perform so collage then, though
it seems quite tricky as internal and unwanted portion
of meshes are constructed.

--

[1]: A Morphological Approach to Curvature-based Evolution of
Curves and Surfaces, Pablo Márquez-Neila, Luis Baumela and Luis Álvarez.
In IEEE Transactions on Pattern Analysis and Machine Intelligence
(PAMI), 2014, DOI 10.1109/TPAMI.2013.106