Add a simple submitter to submit jobs using a file
One of the problem with jobs submission is:

 - the tracking of jobs
 - potential mistakes made on the command line
 - the fact that jobs are run in batch on a set of parameters
 - a description of jobs to make sense of this batch in the history

This introduce a really simple script to submit jobs with a
YAML description of them.

This might help with inspecting results as allows solving the
different problems stated above.

The formatting has to be enforced: this script does not keep
track of eventual mispells on YAML keys, yet.