Adapt mesh decimation to support target face numbers
Previously the mesh decimation step  was using a
simple example from PyMesh.

Now that we use libigl, we can use its decimation method
which allow specifying the target number of faces.

Changes are made to use support this method by default
and to revert to the old one if it fails.